+++ tests/test_direct_upload_limits.py
@@ -44,7 +44,7 @@ def test_direct_upload_routes_use_bounded_reads():
             "read_upload_limited(file, GALLERY_UPLOAD_MAX_BYTES",
             "read_upload_limited(file, GALLERY_TRANSFORM_UPLOAD_MAX_BYTES",
         ],
-        "routes/memory_routes.py": [
+        "routes/memory/memory_routes.py": [
             "read_upload_limited(file, MEMORY_IMPORT_MAX_BYTES",
         ],
         "routes/calendar_routes.py": [
