+++ tests/test_upload_limits_centralized.py
@@ -84,7 +84,7 @@ def test_routes_import_from_upload_limits_not_local_defs():
             'int(os.getenv("ODYSSEUS_GALLERY_UPLOAD_MAX_BYTES"',
             'int(os.getenv("ODYSSEUS_GALLERY_TRANSFORM_UPLOAD_MAX_BYTES"',
         ],
-        "routes/memory_routes.py": ['int(os.getenv("ODYSSEUS_MEMORY_IMPORT_MAX_BYTES"'],
+        "routes/memory/memory_routes.py": ['int(os.getenv("ODYSSEUS_MEMORY_IMPORT_MAX_BYTES"'],
         "routes/personal_routes.py": ['os.getenv("ODYSSEUS_PERSONAL_UPLOAD_MAX_BYTES"'],
         "routes/email_routes.py": ["EMAIL_COMPOSE_UPLOAD_MAX_BYTES = 25 * 1024 * 1024"],
         "routes/stt_routes.py": ["STT_MAX_AUDIO_BYTES = 25 * 1024 * 1024"],
@@ -98,7 +98,7 @@ def test_routes_import_from_upload_limits_not_local_defs():
     # And each imports from upload_limits.
     imports = {
         "routes/gallery/gallery_routes.py": "GALLERY_UPLOAD_MAX_BYTES",
-        "routes/memory_routes.py": "MEMORY_IMPORT_MAX_BYTES",
+        "routes/memory/memory_routes.py": "MEMORY_IMPORT_MAX_BYTES",
         "routes/personal_routes.py": "PERSONAL_UPLOAD_MAX_BYTES",
         "routes/email_routes.py": "EMAIL_COMPOSE_UPLOAD_MAX_BYTES",
         "routes/stt_routes.py": "STT_MAX_AUDIO_BYTES",
