+++ tests/test_vision_owner_scope.py
@@ -90,7 +90,7 @@ def test_request_vision_call_sites_pass_owner():
     upload_source = (ROOT / "routes" / "upload_routes.py").read_text()
     document_source = (ROOT / "routes" / "document_routes.py").read_text()
     gallery_source = (ROOT / "routes" / "gallery" / "gallery_routes.py").read_text()
-    memory_source = (ROOT / "routes" / "memory_routes.py").read_text()
+    memory_source = (ROOT / "routes" / "memory" / "memory_routes.py").read_text()
 
     assert 'analyze_image_with_vl_result(file_info["path"], owner=owner)' in chat_source
     assert "analyze_image_with_vl(path, owner=current_user)" in upload_source
